+++ Sources/CodeEditSourceEditor/Documentation.docc/Documentation.md
@@ -4,36 +4,46 @@ A code editor with syntax highlighting powered by tree-sitter.
 
 ## Overview
 
-![logo](codeedittextview-logo)
-
 An Xcode-inspired code editor view written in Swift powered by tree-sitter for [CodeEdit](https://github.com/CodeEditApp/CodeEdit). Features include syntax highlighting (based on the provided theme), code completion, find and replace, text diff, validation, current line highlighting, minimap, inline messages (warnings and errors), bracket matching, and more.
 
-This package includes both `AppKit` and `SwiftUI` components. It also relies on the `CodeEditLanguages` and `Theme` module. 
+This package includes both `AppKit` and `SwiftUI` components. It also relies on the [`CodeEditLanguages`](https://github.com/CodeEditApp/CodeEditLanguages) for optional syntax highlighting using tree-sitter. 
 
-![banner](preview)
+> **CodeEditSourceEditor is currently in development and it is not ready for production use.** <br> Please check back later for updates on this project. Contributors are welcome as we build out the features mentioned above!
 
-## Syntax Highlighting
+## Currently Supported Languages
 
-``CodeEditSourceEditor`` uses `tree-sitter` for syntax highlighting. A list of already supported languages can be found [here](https://github.com/CodeEditApp/CodeEditSourceEditor/issues/15).
-
-New languages need to be added to the [CodeEditLanguages](https://github.com/CodeEditApp/CodeEditLanguages) repo.
+See this issue [CodeEditLanguages#10](https://github.com/CodeEditApp/CodeEditLanguages/issues/10) on `CodeEditLanguages` for more information on supported languages.
 
 ## Dependencies
 
-Special thanks to both [Marcin Krzyzanowski](https://twitter.com/krzyzanowskim) & [Matt Massicotte](https://twitter.com/mattie) for the great work they've done!
+Special thanks to [Matt Massicotte](https://twitter.com/mattie) for the great work he's done!
 
 | Package | Source | Author |
-| - | - | - |
-| `STTextView` | [GitHub](https://github.com/krzyzanowskim/STTextView) | [Marcin Krzyzanowski](https://twitter.com/krzyzanowskim) |
+| :- | :- | :- |
 | `SwiftTreeSitter` | [GitHub](https://github.com/ChimeHQ/SwiftTreeSitter) | [Matt Massicotte](https://twitter.com/mattie) |
+
+## License
+
+Licensed under the [MIT license](https://github.com/CodeEditApp/CodeEdit/blob/main/LICENSE.md).
 
 ## Topics
 
 ### Text View
 
 - ``CodeEditSourceEditor/CodeEditSourceEditor``
-- ``CodeEditSourceEditor/TextViewController``
+- ``TextViewController``
+- ``GutterView``
 
-### Theme
+### Themes
 
 - ``EditorTheme``
+
+### Text Coordinators
+
+- <doc:TextViewCoordinators>
+- ``TextViewCoordinator``
+- ``CombineCoordinator`` 
+
+### Cursors
+
+- ``CursorPosition``
